More back over to Doubles. Apparently theyre the mack

+++ app/src/BoundingBox.hs
@@ -26,11 +26,11 @@ enlargeBoundingBox :: Position -> AABB -> AABB
 enlargeBoundingBox pos(boxMin, boxMax) = (Vector.min boxMin pos, Vector.max boxMax pos)
 
 -- Linearly scale a box
-scaleBoundingBox :: AABB -> Float -> AABB
+scaleBoundingBox :: AABB -> Double -> AABB
 scaleBoundingBox (boxMin, boxMax) k = (setWTo1 $ boxMin <*> k, setWTo1 $ boxMax <*> k)
 
 -- Give a bounding box a buffer of a certain distance all the way around
-growBoundingBox :: AABB -> Float -> AABB
+growBoundingBox :: AABB -> Double -> AABB
 growBoundingBox (Vector x1 y1 z1 _, Vector x2 y2 z2 _) k = (Vector (x1 - k) (y1 - k) (z1 - k) 1, Vector (x2 + k) (y2 + k) (z2 + k) 1)
 
 -- This is a dummy box that is used initially. If anything is intersected with it, it becomes valid. Else it is an invalid box that can be tested for
@@ -41,10 +41,10 @@ initialInvalidBox = (Vector bigNumber bigNumber bigNumber 1, Vector smallNumber 
       smallNumber = -10000000
 
 -- These functions are useful for finding the greatest or smallest part of a box relative to a normal
-selectMinBoxComponent :: (Vector -> Float) -> Vector -> AABB -> Float
+selectMinBoxComponent :: (Vector -> Double) -> Vector -> AABB -> Double
 selectMinBoxComponent f norm (boxMin, boxMax) = if f norm > 0 then f boxMin else f boxMax
 
-selectMaxBoxComponent :: (Vector -> Float) -> Vector -> AABB -> Float
+selectMaxBoxComponent :: (Vector -> Double) -> Vector -> AABB -> Double
 selectMaxBoxComponent f norm (boxMin, boxMax) = if f norm > 0 then f boxMax else f boxMin
 
 -- Does a box contain a point?
@@ -53,7 +53,7 @@ contains (boxMin, boxMax) p = all insideInterval [vecX, vecY, vecZ]
     where 
       insideInterval f = f p >= f boxMin && f p <= f boxMax
 
-overlapsSphere :: AABB -> Position -> Float -> Bool
+overlapsSphere :: AABB -> Position -> Double -> Bool
 overlapsSphere (boxMin, boxMax) p r = all insideInterval [vecX, vecY, vecZ]
     where 
       insideInterval f = f p >= (f boxMin - r) && f p <= (f boxMax + r)
